% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ASpliSE}
\alias{ASpliSE}
\title{saseRExample}
\format{
saseRExample
}
\source{
saseRExample
}
\usage{
data(saseRExample)
}
\value{
Loads all data needed to run vignette and examples
}
\description{
saseRExample
}
\references{
aa
}
