% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saseR-package.R
\docType{package}
\name{saseR-package}
\alias{saseR}
\alias{saseR-package}
\title{saseR: Scalable Aberrant Splicing and Expression Retrieval}
\description{
saseR is a highly performant and fast framework for aberrant expression and splicing analyses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/statOmics/saseR}
}

}
\author{
\strong{Maintainer}: Alexandre Segers \email{alexandre.segers@ugent.be}

Other contributors:
\itemize{
  \item Jeroen Gilis \email{Jeroen.Gilis@ugent.be} [contributor]
  \item Mattias Van Heetvelde \email{Mattias.Vanheetvelde@ugent.be} [contributor]
  \item Elfride De Baere \email{Elfride.Debaere@ugent.be} [contributor]
  \item Lieven Clement \email{Lieven.Clement@ugent.be} [contributor]
}

}
\keyword{internal}
