% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{saseRExample}
\alias{saseRExample}
\title{saseRExample}
\usage{
data(saseRExample)
}
\value{
Loads all data needed to run examples
}
\description{
saseR example data, obtained from the ASpli package
(https://bioconductor.org/packages/release/bioc/html/ASpli.html,
10.1093/bioinformatics/btab141).
}
