% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{SEjunctions}
\alias{SEjunctions}
\title{SEjunctions}
\usage{
data(saseRExample)
}
\value{
Loads all data needed to run vignette and examples
}
\description{
SummarizedExperiment containing junction-level counts to
perform an example aberrant splicing analysis
}
