% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ASpliSE}
\alias{ASpliSE}
\title{ASpliSE}
\usage{
data(saseRExample)
}
\value{
Loads all data needed to run vignette and examples
}
\description{
SummarizedExperiment containing gene counts, bin counts and junction counts in the metadata slots, together with some rowData information. Output from BamtoAspliCounts.
}
