% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertASpli.R
\name{convertASpli}
\alias{convertASpli}
\title{Converting SummarizedExperiment with different ASpli-counts to
gene, bin or junction level counts.}
\usage{
convertASpli(ASpliSE, type = "none", filter = TRUE, ...)
}
\arguments{
\item{ASpliSE}{An SummarizedExperiment object obtained by the
`BamtoAspliCounts` function, which contains gene, bin and junction level
counts in the metadata slot.}

\item{type}{character vector specifying the counts to be extracted. Can be
`gene`, `bin` and `junction`.}

\item{filter}{logical value specifying to filter bin counts based on the
type of bin. Regions that contain both intronic and exonic regions are
filtered as the intron and exon bins are already defined in other features.}
}
\value{
A `SummarizedExperiment` instance, representing the `gene`, `bin` or
`junction` counts as specified.
}
\description{
convertASpli is used to obtain gene, bin or junction level counts from
a SummarizedExperiment object containing these three in the metadata slot.
}
\examples{

CountMatrix <- matrix(rnbinom(n = 1000, mu = 10, size = 5), ncol = 20, nrow = 50)
Metadata <- c(rep(c("Male","Female"), each = 10))
se <- SummarizedExperiment::SummarizedExperiment(
assays = list("counts" = CountMatrix),
 colData = data.frame("Sex" = Metadata),
 metadata = list(design = ~Sex))
se <- findEncodingDimRUV(se = se, dimensions = c(0:10))
se <- fitRUVMD(se = se)
SummarizedExperiment::assays(se)[["mu"]][1:5,1:5]
SummarizedExperiment::assays(se)[["pValue"]][1:5,1:5]
SummarizedExperiment::assays(se)[["pValueAdjust"]][1:5,1:5]
SummarizedExperiment::rowData(se)[["theta"]][1:5]
S4Vectors::metadata(se)[["optimalEncDim"]]

}
