% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{features}
\alias{features}
\title{features}
\format{
features
}
\source{
features
}
\usage{
data(saseRExample)
}
\value{
Loads all data needed to run vignette and examples
}
\description{
features
}
\references{
aa
}
