% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{SEbins}
\alias{SEbins}
\title{SEbins}
\usage{
data(saseRExample)
}
\value{
Loads all data needed to run vignette and examples
}
\description{
SummarizedExperiment containing bin-level counts to perform
an example aberrant splicing analysis.
}
