% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{SEgenes}
\alias{SEgenes}
\title{SEgenes}
\usage{
data(saseRExample)
}
\value{
Loads all data needed to run vignette and examples
}
\description{
SummarizedExperiment containing gene-level counts to perform
an example aberrant expression analysis.
}
