% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{SEjunctions}
\alias{SEjunctions}
\title{SEjunctions}
\format{
SEjunctions
}
\source{
SEjunctions
}
\usage{
data(saseRExample)
}
\value{
Loads all data needed to run vignette and examples
}
\description{
SEjunctions
}
\references{
SEjunctions
}
